requirements.txt
anthropic>=0.34.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
apscheduler>=3.10.4
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0
beautifulsoup4>=4.12.3
lxml>=5.2.0
requests>=2.31.0
pydantic>=2.7.0
